Fix 64 bits PCI memory BARs not supported

+++ interfaces/pci/src/ffi.rs
@@ -153,7 +153,7 @@ pub struct PciDeviceInfo {
 // TODO: actually figure out PCI and adjust this
 #[derive(Debug, Clone, Encode, Decode)]
 pub enum PciBaseAddressRegister {
-    Memory { base_address: u32 },
+    Memory { base_address: u64 },
     Io { base_address: u32 },
 }
